requirements.txt
aioboto3==13.2.0
aiohttp==3.9.5
fitz==0.0.1.dev2
graspologic==3.4.1
lmdeploy==0.6.4
nano_vectordb==0.0.4.3
networkx==3.2.1
numpy==2.2.0
ollama==0.4.4
openai==1.57.3
pandas==2.2.3
Pillow==11.0.0
protobuf==5.29.1
pydantic==2.10.3
qwen_vl_utils==0.0.8
tables==3.10.1
tenacity==8.3.0
tiktoken==0.8.0
torch==2.0.1
tqdm==4.66.4
transformers==4.45.2
google-generativeai
openai
